# openapi/openapi.snapshot.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Limiq.io API",
    "summary": "Identity & Permission Layer for Autonomous Agents",
    "description": "\n## Limiq.io API\n\nIdentity & permission layer for autonomous agents.\n\n### Auth model (MVP)\nSensitive routes require `X-Workspace-Id` and `X-Workspace-Key`. The key is returned\nonce by workspace bootstrap and is verified in constant time. The request\n`workspace_id` must match the authenticated workspace.\n\n`POST /workspaces` is a bootstrap route protected by `X-Bootstrap-Token`.\n\n### Error format\nBusiness errors are returned as:\n\n```json\n{\"detail\": {\"code\": \"SOME_CODE\", \"message\": \"Human readable message\"}}\n```\n\n### Main flow\n1. Create agent\n2. Create policy\n3. Bind policy to agent\n4. Request capability\n5. Verify signed action\n6. Query/export audit logs\n7. Check audit chain integrity\n",
    "termsOfService": "https://github.com/qurveai/limiq.io",
    "contact": {
      "name": "Limiq.io Maintainers",
      "url": "https://github.com/qurveai/limiq.io"
    },
    "license": {
      "name": "Apache-2.0"
    },
    "version": "0.5.0",
    "x-logo": {
      "url": "https://raw.githubusercontent.com/redocly/redoc/main/demo/logo.png",
      "altText": "Limiq.io"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local development"
    }
  ],
  "paths": {
...
    "/audit/events": {
      "get": {
        "tags": [
          "audit"
        ],
        "summary": "List Audit Events",
        "description": "Returns paginated audit events with optional filters.",
        "operationId": "list_audit_events_endpoint_audit_events_get",
        "security": [
          {
            "APIKeyHeader": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "workspace_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Workspace identifier",
              "title": "Workspace Id"
            },
            "description": "Workspace identifier"
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Start datetime",
              "title": "From"
            },
            "description": "Start datetime"
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "End datetime",
              "title": "To"
            },
            "description": "End datetime"
          },
          {
            "name": "event_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by event type",
              "title": "Event Type"
            },
            "description": "Filter by event type"
          },
          {
            "name": "subject_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by subject id",
              "title": "Subject Id"
            },
            "description": "Filter by subject id"
          },
          {
            "name": "decision",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "ALLOW",
                    "DENY"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ALLOW or DENY filter",
              "title": "Decision"
            },
            "description": "ALLOW or DENY filter"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "Page size",
              "default": 50,
              "title": "Limit"
            },
            "description": "Page size"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Page offset",
              "default": 0,
              "title": "Offset"
            },
            "description": "Page offset"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditEventsListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication headers.",
            "content": {
              "application/json": {
                "example": {
                  "detail": {
                    "code": "AUTH_WORKSPACE_KEY_MISSING",
                    "message": "Missing X-Workspace-Key header"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Workspace mismatch with authenticated context.",
            "content": {
              "application/json": {
                "example": {
                  "detail": {
                    "code": "WORKSPACE_MISMATCH",
                    "message": "Workspace does not match authenticated context"
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation error on payload/query params.",
            "content": {
              "application/json": {
                "example": {
                  "detail": {
                    "code": "VALIDATION_ERROR",
                    "message": "Query param 'from' must be <= 'to'"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/audit/export.json": {
      "get": {
        "tags": [
          "audit"
        ],
        "summary": "Export Audit Events (JSON)",
        "description": "Exports filtered audit events as JSON.",
        "operationId": "export_audit_json_endpoint_audit_export_json_get",
        "security": [
          {
            "APIKeyHeader": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "workspace_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Workspace identifier",
              "title": "Workspace Id"
            },
            "description": "Workspace identifier"
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Start datetime",
              "title": "From"
            },
            "description": "Start datetime"
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "End datetime",
              "title": "To"
            },
            "description": "End datetime"
          },
          {
            "name": "event_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by event type",
              "title": "Event Type"
            },
            "description": "Filter by event type"
          },
          {
            "name": "subject_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by subject id",
              "title": "Subject Id"
            },
            "description": "Filter by subject id"
          },
          {
            "name": "decision",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "ALLOW",
                    "DENY"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ALLOW or DENY filter",
              "title": "Decision"
            },
            "description": "ALLOW or DENY filter"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuditEventResponse"
                  },
                  "title": "Response Export Audit Json Endpoint Audit Export Json Get"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication headers.",
            "content": {
              "application/json": {
                "example": {
                  "detail": {
                    "code": "AUTH_WORKSPACE_KEY_MISSING",
                    "message": "Missing X-Workspace-Key header"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Workspace mismatch with authenticated context.",
            "content": {
              "application/json": {
                "example": {
                  "detail": {
                    "code": "WORKSPACE_MISMATCH",
                    "message": "Workspace does not match authenticated context"
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation error on payload/query params.",
            "content": {
              "application/json": {
                "example": {
                  "detail": {
                    "code": "VALIDATION_ERROR",
                    "message": "Query param 'from' must be <= 'to'"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/audit/export.csv": {
      "get": {
        "tags": [
          "audit"
        ],
        "summary": "Export Audit Events (CSV)",
        "description": "Exports filtered audit events as CSV.",
        "operationId": "export_audit_csv_endpoint_audit_export_csv_get",
        "security": [
          {
            "APIKeyHeader": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "workspace_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Workspace identifier",
              "title": "Workspace Id"
            },
            "description": "Workspace identifier"
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Start datetime",
              "title": "From"
            },
            "description": "Start datetime"
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "End datetime",
              "title": "To"
            },
            "description": "End datetime"
          },
          {
            "name": "event_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by event type",
              "title": "Event Type"
            },
            "description": "Filter by event type"
          },
          {
            "name": "subject_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by subject id",
              "title": "Subject Id"
            },
            "description": "Filter by subject id"
          },
          {
            "name": "decision",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "ALLOW",
                    "DENY"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ALLOW or DENY filter",
              "title": "Decision"
            },
            "description": "ALLOW or DENY filter"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication headers.",
            "content": {
              "application/json": {
                "example": {
                  "detail": {
                    "code": "AUTH_WORKSPACE_KEY_MISSING",
                    "message": "Missing X-Workspace-Key header"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Workspace mismatch with authenticated context.",
            "content": {
              "application/json": {
                "example": {
                  "detail": {
                    "code": "WORKSPACE_MISMATCH",
                    "message": "Workspace does not match authenticated context"
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation error on payload/query params.",
            "content": {
              "application/json": {
                "example": {
                  "detail": {
                    "code": "VALIDATION_ERROR",
                    "message": "Query param 'from' must be <= 'to'"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/audit/integrity/check": {
      "get": {
        "tags": [
          "audit"
        ],
        "summary": "Check Audit Chain Integrity",
        "description": "Verifies hash-chain continuity in a workspace. Returns OK, BROKEN or PARTIAL.",
        "operationId": "check_audit_integrity_endpoint_audit_integrity_check_get",
        "security": [
          {
            "APIKeyHeader": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "workspace_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Workspace identifier",
              "title": "Workspace Id"
            },
            "description": "Workspace identifier"
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Start datetime",
              "title": "From"
            },
            "description": "Start datetime"
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "End datetime",
              "title": "To"
            },
            "description": "End datetime"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditIntegrityResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication headers.",
            "content": {
              "application/json": {
                "example": {
                  "detail": {
                    "code": "AUTH_WORKSPACE_KEY_MISSING",
                    "message": "Missing X-Workspace-Key header"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Workspace mismatch with authenticated context.",
            "content": {
              "application/json": {
                "example": {
                  "detail": {
                    "code": "WORKSPACE_MISMATCH",
                    "message": "Workspace does not match authenticated context"
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation error on payload/query params.",
            "content": {
              "application/json": {
                "example": {
                  "detail": {
                    "code": "VALIDATION_ERROR",
                    "message": "Query param 'from' must be <= 'to'"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuditEventResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "event_time": {
            "type": "string",
            "format": "date-time",
            "title": "Event Time"
          },
          "event_type": {
            "type": "string",
            "title": "Event Type"
          },
          "workspace_id": {
            "type": "string",
            "format": "uuid",
            "title": "Workspace Id"
          },
          "actor_type": {
            "type": "string",
            "title": "Actor Type"
          },
          "actor_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actor Id"
          },
          "subject_type": {
            "type": "string",
            "title": "Subject Type"
          },
          "subject_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject Id"
          },
          "event_data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Event Data"
          },
          "prev_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prev Hash"
          },
          "event_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Hash"
          }
        },
        "type": "object",
        "required": [
          "id",
          "event_time",
          "event_type",
          "workspace_id",
          "actor_type",
          "actor_id",
          "subject_type",
          "subject_id",
          "event_data",
          "prev_hash",
          "event_hash"
        ],
        "title": "AuditEventResponse"
      },
      "AuditEventsListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AuditEventResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "count",
          "limit",
          "offset"
        ],
        "title": "AuditEventsListResponse"
      },
      "AuditIntegrityResponse": {
        "properties": {
          "workspace_id": {
            "type": "string",
            "format": "uuid",
            "title": "Workspace Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "OK",
              "BROKEN",
              "PARTIAL"
            ],
            "title": "Status"
          },
          "checked_count": {
            "type": "integer",
            "title": "Checked Count"
          },
          "broken_at_event_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Broken At Event Id"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "workspace_id",
          "status",
          "checked_count",
          "broken_at_event_id",
          "message"
        ],
        "title": "AuditIntegrityResponse"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Workspace-Key"
      }
    }
  },
...
}